# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api documentation",
        "description": "Api documentation",
        "contact": {
            "email": "saguilarsanhuza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener una lista de productos",
                "operationId": "indexProducts",
                "responses": {
                    "200": {
                        "description": "Una lista de tiendas de los productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stores": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar un  producto",
                "operationId": "putProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "price",
                                    "old_price",
                                    "discount",
                                    "stock_product",
                                    "product_type_id"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Bebida más jugo"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "2500"
                                    },
                                    "discount": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "20"
                                    },
                                    "stock_product": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "2"
                                    },
                                    "product_type_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mensaje de exito al editar el producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto editar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Mesaje de error en caso de que el producto no es editado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error  editar producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "price",
                                    "old_price",
                                    "discount",
                                    "stock_product",
                                    "product_type_id"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Bebida más jugo"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "2500"
                                    },
                                    "discount": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "20"
                                    },
                                    "stock_product": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "2"
                                    },
                                    "product_type_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mensaje de exito al crear el producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Mesaje de error en caso de que el producto no es creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error crear producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Recupera un producto",
                "description": "Recupera un producto",
                "operationId": "showProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto recuperada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar un producto",
                "operationId": "destroyProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Mesaje de producto eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prducto eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Mesanje de error al eliminar el producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al eliminar producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/store/{store}": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Recupera una tienda y sus productos",
                "description": "Recupera una tienda y sus productos",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tienda recuperada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "store": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tienda no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tienda no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Stores"
                ],
                "summary": "Actualizar tiendas",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "store",
                        "in": "path",
                        "description": "El ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image_store",
                                    "direction",
                                    "delivery_type_id",
                                    "actual_price",
                                    "old_price",
                                    "opening_hours",
                                    "closing_time",
                                    "logo_store"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Oxxo"
                                    },
                                    "image_store": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "/storage/images/image.jpg"
                                    },
                                    "direction": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "av los toros"
                                    },
                                    "delivery_type_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "actual_price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "25000"
                                    },
                                    "old_price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "25000"
                                    },
                                    "opening_hours": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "9:00"
                                    },
                                    "closing_time": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "21:00"
                                    },
                                    "logo_store": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "/storage/images/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tienda actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tienda editada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar tienda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al editar tienda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Eliminar una tienda",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "store",
                        "in": "path",
                        "description": "El ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tienda eliminada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tienda eliminada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar la tienda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al eliminar tienda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}